fix: Remove magic number that used to 'fix' arrow head positions
This was a hack that I could not really explain.
It looks like in the end this was due to the fact that we
computed the position of an artificially added element and not the
actual provided element.

The hack is no longer necessary and arrow heads should be pixel perfect now!